split the single current test file into two - one having all the tests relevant only for the `iterified` function and the other all the tests relevant only for `iterifiedUnwrapped`
